requirements.txt
PyQt6>=6.5
google-cloud-pubsub>=2.20.0
google-auth>=2.0.0
